- Floatify `InputPacket::fvel` and `InputPacket::svel`.
* Since all velocities are floated, we can now take each game's keymove scaling into its ticker and use a consistent keymove size, making non-mouselook and controller input truly identical across the games without having to try and fudge any values.

+++ source/core/packet.h
@@ -69,8 +69,8 @@ enum
 
 struct InputPacket
 {
-    int16_t svel;
-    int16_t fvel;
+    float svel;
+    float fvel;
     float avel;
     float horz;
     ESyncBits actions;
